+++ test/draft-test.js
@@ -722,7 +722,9 @@ describe("draft", () => {
     });
     response = await util.callRead(
       request,
-      decodeURIComponent(`/v2/draft/Header(ID=guid'${id}',IsActiveEntity=false)/Items?$filter=((IsActiveEntity eq true and substringof('Item2',name)) or (IsActiveEntity eq false and (substringof('Item2',name) or HasActiveEntity eq false))) or description eq 'ABC'`)
+      decodeURIComponent(
+        `/v2/draft/Header(ID=guid'${id}',IsActiveEntity=false)/Items?$filter=((IsActiveEntity eq true and substringof('Item2',name)) or (IsActiveEntity eq false and (substringof('Item2',name) or HasActiveEntity eq false))) or description eq 'ABC'`
+      )
     );
     expect(response.body).toBeDefined();
     expect(response.body.d).toBeDefined();
@@ -744,7 +746,9 @@ describe("draft", () => {
     });
     response = await util.callRead(
       request,
-      decodeURIComponent(`/v2/draft/Header(ID=guid'${id}',IsActiveEntity=false)/Items?$filter=(IsActiveEntity eq true and name eq 'Test Item') or (IsActiveEntity eq false and (name eq 'Test Item' or HasActiveEntity eq false))`)
+      decodeURIComponent(
+        `/v2/draft/Header(ID=guid'${id}',IsActiveEntity=false)/Items?$filter=(IsActiveEntity eq true and name eq 'Test Item') or (IsActiveEntity eq false and (name eq 'Test Item' or HasActiveEntity eq false))`
+      )
     );
     expect(response.body).toBeDefined();
     expect(response.body.d).toBeDefined();